requirements.txt

adjustText==0.8
albumentations==1.3.0
numpy==1.21.6
matplotlib==3.5.3
umap-learn==0.5.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
scipy==1.5.0
scikit-image==0.19.3
scikit-learn==1.0.2
seaborn==0.12.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.1
pandas==1.3.5
PyYAML==5.3.1
jupyter==1.0.0
